refactor!: Make Parameters clone
BREAKING CHANGE: To do this, I removed the seed string from the struct.
It turns out that not having this does not affect any of the test
results. I will be going through and removing references to the seed in
subsequent PRs, as it does not seem to be important.